+++ integreat_cms/static/src/js/collapsible-boxes.ts
@@ -2,9 +2,12 @@
  * Handles toggling boxes on the dashboard
  * @param {Event} event Click event for dashboard containers
  */
+
+import { getCookie } from "./utils/cookies";
+
 const toggleDashboardSection = async ({ target }: Event): Promise<void> => {
     // Toggle arrows of the clicked widget
-    const collapsible = (target as HTMLElement).closest(".collapsible");
+    const collapsible = (target as HTMLElement).closest(".collapsible") as HTMLElement;
     collapsible.querySelector(".up-arrow").classList.toggle("hidden");
     collapsible.querySelector(".down-arrow").classList.toggle("hidden");
     // Toggle the content div which should be collapsed or expanded
@@ -26,6 +29,7 @@ const toggleDashboardSection = async ({ target }: Event): Promise<void> => {
         setTimeout(() => {
             content.style.height = "auto";
         }, timeoutDuration);
+        collapsible.removeAttribute("data-collapsed");
     } else {
         // Set height to explicit numerical value instead of "auto" to enable continuous transition
         content.style.height = `${content.clientHeight}px`;
@@ -33,11 +37,32 @@ const toggleDashboardSection = async ({ target }: Event): Promise<void> => {
         setTimeout(() => {
             content.style.height = "0px";
         });
+        collapsible.dataset.collapsed = "";
     }
+    // Get the ids of all boxes which are collapsed and have a non-empty id attribute
+    const collapsedBoxes = Array.from(document.querySelectorAll("[data-collapsed]:not([id=''])")).map((box) => box.id);
+    // Set the scope of the cookie to the first two path segments (slice to 3 because the pathname starts with a leading slash)
+    const sliceEnd = 3;
+    const cookiePath = window.location.pathname.split("/").slice(0, sliceEnd).join("/");
+    // Store the currently collapsed boxes in a cookie
+    document.cookie = `collapsed-boxes=${JSON.stringify(collapsedBoxes)};path=${cookiePath}`;
 };
 
 window.addEventListener("load", () => {
     document.querySelectorAll(".collapsible").forEach((node) => {
         node.addEventListener("click", toggleDashboardSection);
     });
+    // Restore all collapsed boxes
+    const collapsedBoxes = getCookie("collapsed-boxes");
+    if (collapsedBoxes) {
+        JSON.parse(collapsedBoxes).forEach((id: string) => {
+            const collapsible = document.querySelector(`.collapsible[id='${id}']`) as HTMLElement;
+            if (collapsible) {
+                collapsible.querySelector(".up-arrow").classList.add("hidden");
+                collapsible.querySelector(".down-arrow").classList.remove("hidden");
+                collapsible.parentNode.querySelector(".collapsible-content").classList.add("h-0");
+                collapsible.dataset.collapsed = "";
+            }
+        });
+    }
 });